refactor: split infrastructure into primary and secondary
