+++ __tests__/integration/cache.service.spec.ts
@@ -1,5 +1,5 @@
 import { ctx } from '__tests__/setup-integration-context.js';
-import { CacheProvider } from 'src/infrastructure/adapters/cache.provider.js';
+import { CacheProvider } from 'src/secondary/adapters/cache.provider.js';
 import { scheduler } from 'node:timers/promises';
 
 describe('CacheService (integration)', () => {
