+++ __tests__/integration/metrics.spec.ts
@@ -1,7 +1,7 @@
 import { ctx } from '__tests__/setup-integration-context.js';
-import { HttpStatus } from 'src/infrastructure/types/http.types.js';
-import { Server } from 'src/infrastructure/server.js';
-import { MetricsProvider } from 'src/infrastructure/adapters/metrics.provider.js';
+import { HttpStatus } from 'src/secondary/types/http.types.js';
+import { Server } from 'src/primary/server.js';
+import { MetricsProvider } from 'src/secondary/adapters/metrics.provider.js';
 
 describe('metrics controller (integration)', () => {
   let server: Server;
