+++ __tests__/integration/weather.spec.ts
@@ -1,8 +1,8 @@
 import { ctx } from '__tests__/setup-integration-context.js';
-import { HttpStatus } from 'src/infrastructure/types/http.types.js';
+import { HttpStatus } from 'src/secondary/types/http.types.js';
 import { MockInstance } from 'vitest';
 import { Exception } from 'src/domain/entities/exception.entity.js';
-import { Server } from 'src/infrastructure/server.js';
+import { Server } from 'src/primary/server.js';
 import { WeatherService } from 'src/domain/services/weather.service.js';
 
 describe('weather controller (integration)', () => {
