+++ src/index.tsx
@@ -1,4 +1,4 @@
-import { Container } from './infrastructure/container.js';
+import { Container } from './primary/adapters/container.js';
 
 const main = async () => {
   const { app } = new Container();
