+++ src/primary/controllers/weather.controller.ts
