+++ src/secondary/adapters/weather/open-meteo.provider.ts
