+++ src/secondary/templates/weather-update.template.tsx
